+++ api/rust/crates/lief/tests/coff_tests.rs
@@ -18,6 +18,7 @@ fn explore_coff(bin_name: &str, coff: &lief::coff::Binary) {
     for section in coff.sections() {
         format!("{section:?} {section}");
         format!("nb relocs: {}", section.relocations().len());
+        assert!(coff.section_by_name(&section.name()).is_some());
         for symbol in section.symbols() {
             format!("{symbol:?} {symbol}");
         }
@@ -52,6 +53,7 @@ fn explore_coff(bin_name: &str, coff: &lief::coff::Binary) {
 
     assert!(coff.find_string(0).is_none());
     assert!(coff.find_string(4).is_some());
+    assert!(coff.section_by_name("<not-a-section>").is_none());
 }
 
 fn test_with(bin_name: &str) {
@@ -76,4 +78,29 @@ fn test_api() {
     test_with("x64_debug_cl.obj");
     test_with("arm64_debug_cl_bigobj.obj");
     test_with("arm64_debug_cl.obj");
+}
+
+#[test]
+fn test_section_by_name() {
+    let path = utils::get_coff_sample("arm64_debug_cl.obj").unwrap();
+    let coff = lief::coff::Binary::parse(path.to_str().unwrap()).unwrap();
+
+    assert!(coff.section_by_name("<not-a-section>").is_none());
+
+    let section = coff.section_by_name(".drectve").unwrap();
+    assert_eq!(section.name(), ".drectve");
+    assert_eq!(section.sizeof_raw_data(), 0x91);
+    assert!(section.coff_string().is_none());
+
+    let path = utils::get_coff_sample("dwarf.obj").unwrap();
+    let coff = lief::coff::Binary::parse(path.to_str().unwrap()).unwrap();
+
+    let section = coff.section_by_name(".debug_rnglists").unwrap();
+    assert_eq!(section.name(), "/18");
+    assert_eq!(section.coff_string().unwrap().str(), ".debug_rnglists");
+
+    let section = coff.section_by_name("/18").unwrap();
+    assert_eq!(section.coff_string().unwrap().str(), ".debug_rnglists");
+
+    assert!(coff.section_by_name(".debug_line").is_none());
 }